Add stat_custom_receive0_management_mode to the Broadcom config file of TH5 platforms
<!--
     Please make sure you've read and understood our contributing guidelines:
     https://github.com/Azure/SONiC/blob/gh-pages/CONTRIBUTING.md

     ** Make sure all your commits include a signature generated with `git commit -s` **

     If this is a bug fix, make sure your description includes "fixes #xxxx", or
     "closes #xxxx" or "resolves #xxxx"

     Please provide the following information:
-->

#### Why I did it

There's an upgrade in SAI / SDK so that we need to add `stat_custom_receive0_management_mode` to config files of TH5 products. Plesae see Broadcom article "SDK - API Behavior Change - Statistics counter: snmpBcmCustomReceive0 is reserved" (ID: 26564) for more details.

Without this change, we'll see `orchagent` crashing with following errors:
```
2025 May 30 17:06:21.814814 moby102 ERR syncd#syncd: [none] SAI_API_SWITCH:_brcm_sai_stats_init:19994 Custom stat RFILDR delete failed with error Invalid configuration (0xfffffff1).
2025 May 30 17:06:21.814814 moby102 CRIT syncd#syncd: [none] SAI_API_SWITCH:brcm_sai_create_switch:4202 configuring stats failed with error -1.
2025 May 30 17:06:21.814814 moby102 ERR syncd#syncd: :- sendApiResponse: api SAI_COMMON_API_CREATE failed in syncd mode: SAI_STATUS_FAILURE
```

##### Work item tracking
- Microsoft ADO **(number only)**:

#### How I did it

Update the Broadcom config files based on the suggestion provided by Broadcom

#### How to verify it

We verified the crash is fixed with this patch, and running `sonic-mgmt` with this patch.

<!--
If PR needs to be backported, then the PR must be tested against the base branch and the earliest backport release branch and provide tested image version on these two branches. For example, if the PR is requested for master, 202211 and 202012, then the requester needs to provide test results on master and 202012.
-->

#### Which release branch to backport (provide reason below if selected)

<!--
- Note we only backport fixes to a release branch, *not* features!
- Please also provide a reason for the backporting below.
- e.g.
- [x] 202006
-->

- [ ] 202205
- [ ] 202211
- [ ] 202305
- [ ] 202311
- [ ] 202405
- [ ] 202411
- [x] 202505
- [x] msft-202412
- [x] msft-202503

#### Tested branch (Please provide the tested image version)

<!--
- Please provide tested image version
- e.g.
- [x] 20201231.100
-->

- [ ] <!-- image version 1 -->
- [ ] <!-- image version 2 -->
- [x] msft-202412

#### Description for the changelog
<!--
Write a short (one line) summary that describes the changes in this
pull request for inclusion in the changelog:
-->

<!--
 Ensure to add label/tag for the feature raised. example - PR#2174 under sonic-utilities repo. where, Generic Config and Update feature has been labelled as GCU.
-->

#### Link to config_db schema for YANG module changes
<!--
Provide a link to config_db schema for the table for which YANG model
is defined
Link should point to correct section on https://github.com/Azure/sonic-buildimage/blob/master/src/sonic-yang-models/doc/Configuration.md
-->

#### A picture of a cute animal (not mandatory but encouraged)